refactor: update the metadadata classes to use a close nameclature
