+++ tck/src/main/resources/ee/jakarta/tck/data/framework/signature/jakarta.data.sig_17
@@ -93,18 +93,18 @@ hfds serialVersionUID
 
 CLSS abstract interface jakarta.data.exceptions.package-info
 
-CLSS public jakarta.data.metamodel.Attribute
-meth public final boolean init(jakarta.data.metamodel.AttributeInfo)
+CLSS public jakarta.data.metamodel.AttributeMetadata
+meth public final boolean init(jakarta.data.metamodel.Attribute)
 meth public final jakarta.data.Sort asc()
 meth public final jakarta.data.Sort ascIgnoreCase()
 meth public final jakarta.data.Sort desc()
 meth public final jakarta.data.Sort descIgnoreCase()
 meth public final java.lang.String name()
-meth public final static jakarta.data.metamodel.Attribute get()
+meth public final static jakarta.data.metamodel.AttributeMetadata get()
 supr java.lang.Object
 hfds info
 
-CLSS public abstract interface jakarta.data.metamodel.AttributeInfo
+CLSS public abstract interface jakarta.data.metamodel.Attribute
 meth public abstract jakarta.data.Sort asc()
 meth public abstract jakarta.data.Sort ascIgnoreCase()
 meth public abstract jakarta.data.Sort desc()
